Drop support for old Symfony versions
