+++ composer.json
@@ -20,15 +20,18 @@
     ],
     "require": {
         "symfony/framework-bundle": "^2.3.3|^3.0",
-        "symfony/console": "^2.3.3|^3.0",
-        "symfony/validator": "^2.3.20|^3.0",
-        "symfony/polyfill-mbstring": "^1.3.0"
         "php": "^5.6 || ^7.0",
+        "symfony/validator": "^3.3.6",
+        "symfony/polyfill-mbstring": "^1.5.0"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^3.2.6",
         "matthiasnoback/symfony-dependency-injection-test": "^1.1.0",
         "matthiasnoback/symfony-service-definition-validator": "^1.2.2"
+        "matthiasnoback/symfony-dependency-injection-test": "^2.1.0",
+        "matthiasnoback/symfony-service-definition-validator": "^1.2.8",
+        "symfony/console": "^3.3.6",
+        "symfony/phpunit-bridge": "^3.3.6"
     },
     "autoload": {
         "psr-4": {
