+++ src/Validator/Constraints/PasswordRequirementsValidator.php
@@ -13,12 +13,12 @@ namespace Rollerworks\Bundle\PasswordStrengthBundle\Validator\Constraints;
 
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
-use Symfony\Component\Validator\Context\ExecutionContextInterface;
+use Symfony\Component\Validator\Exception\UnexpectedTypeException;
 
 class PasswordRequirementsValidator extends ConstraintValidator
 {
     /**
-     * @param string                          $value
+     * @param null|string                     $value
      * @param PasswordRequirements|Constraint $constraint
      */
     public function validate($value, Constraint $constraint)
@@ -27,55 +27,39 @@ class PasswordRequirementsValidator extends ConstraintValidator
             return;
         }
 
+        if (!is_scalar($value) && !(is_object($value) && method_exists($value, '__toString'))) {
+            throw new UnexpectedTypeException($value, 'string');
+        }
+
         if ($constraint->minLength > 0 && (mb_strlen($value) < $constraint->minLength)) {
-            if ($this->context instanceof ExecutionContextInterface) {
-                $this->context->buildViolation($constraint->tooShortMessage)
-                    ->setParameters(array('{{length}}' => $constraint->minLength))
-                    ->setInvalidValue($value)
-                    ->addViolation();
-            } else {
-                $this->context->addViolation($constraint->tooShortMessage, array('{{length}}' => $constraint->minLength), $value);
-            }
+            $this->context->buildViolation($constraint->tooShortMessage)
+                ->setParameters(array('{{length}}' => $constraint->minLength))
+                ->setInvalidValue($value)
+                ->addViolation();
         }
 
         if ($constraint->requireLetters && !preg_match('/\pL/u', $value)) {
-            if ($this->context instanceof ExecutionContextInterface) {
-                $this->context->buildViolation($constraint->missingLettersMessage)
-                    ->setInvalidValue($value)
-                    ->addViolation();
-            } else {
-                $this->context->addViolation($constraint->missingLettersMessage, array(), $value);
-            }
+            $this->context->buildViolation($constraint->missingLettersMessage)
+                ->setInvalidValue($value)
+                ->addViolation();
         }
 
         if ($constraint->requireCaseDiff && !preg_match('/(\p{Ll}+.*\p{Lu})|(\p{Lu}+.*\p{Ll})/u', $value)) {
-            if ($this->context instanceof ExecutionContextInterface) {
-                $this->context->buildViolation($constraint->requireCaseDiffMessage)
-                    ->setInvalidValue($value)
-                    ->addViolation();
-            } else {
-                $this->context->addViolation($constraint->requireCaseDiffMessage, array(), $value);
-            }
+            $this->context->buildViolation($constraint->requireCaseDiffMessage)
+                ->setInvalidValue($value)
+                ->addViolation();
         }
 
         if ($constraint->requireNumbers && !preg_match('/\pN/u', $value)) {
-            if ($this->context instanceof ExecutionContextInterface) {
-                $this->context->buildViolation($constraint->missingNumbersMessage)
-                    ->setInvalidValue($value)
-                    ->addViolation();
-            } else {
-                $this->context->addViolation($constraint->missingNumbersMessage, array(), $value);
-            }
+            $this->context->buildViolation($constraint->missingNumbersMessage)
+                ->setInvalidValue($value)
+                ->addViolation();
         }
 
         if ($constraint->requireSpecialCharacter && !preg_match('/[^p{Ll}\p{Lu}\pL\pN]/u', $value)) {
-            if ($this->context instanceof ExecutionContextInterface) {
-                $this->context->buildViolation($constraint->missingSpecialCharacterMessage)
-                    ->setInvalidValue($value)
-                    ->addViolation();
-            } else {
-                $this->context->addViolation($constraint->missingSpecialCharacterMessage, array(), $value);
-            }
+            $this->context->buildViolation($constraint->missingSpecialCharacterMessage)
+                ->setInvalidValue($value)
+                ->addViolation();
         }
     }
 }
