+++ tests/Validator/BlacklistValidationTest.php
@@ -39,11 +39,6 @@ class BlacklistValidationTest extends AbstractConstraintValidatorTest
         );
     }
 
-    protected function getApiVersion()
-    {
-        return Validation::API_VERSION_2_5;
-    }
-
     protected function createValidator()
     {
         $provider = new ArrayProvider(array('test', 'foobar'));
