+++ tests/Validator/PasswordRequirementsValidatorTest.php
@@ -38,11 +38,6 @@ class PasswordRequirementsValidatorTest extends AbstractConstraintValidatorTest
         );
     }
 
-    protected function getApiVersion()
-    {
-        return Validation::API_VERSION_2_5;
-    }
-
     protected function createValidator()
     {
         return new PasswordRequirementsValidator();
