+++ tests/Validator/PasswordStrengthTest.php
@@ -50,11 +50,6 @@ class PasswordStrengthTest extends AbstractConstraintValidatorTest
         );
     }
 
-    protected function getApiVersion()
-    {
-        return Validation::API_VERSION_2_5;
-    }
-
     protected function createValidator()
     {
         return new PasswordStrengthValidator(new Translator('en'));
